fix: Set correct Content-Type when serving assets
Users have reported problems using either Chrome, Edge or Safari getting the error message

failed to load module script: Expected a JavaScript module script but the server responded with a MIME type of "application/octet-stream". Strict MIME type checking is enforced for module scripts per HTML spec.

The root cause was that we set the Content-Type header twice, once to application/octet-stream and the second to text/javascript.
With this fix only a single (correct) Content-Type header is sent.